+++ core/src/commonMain/kotlin/nl/adaptivity/xmlutil/EventType.kt
@@ -197,6 +197,8 @@ public enum class EventType {
 
         override val isIgnorable: Boolean get() = true
 
+        override val isTextElement: Boolean get() = true
+
         override fun createEvent(reader: XmlReader): TextEvent =
             TextEvent(reader.locationInfo, PROCESSING_INSTRUCTION, reader.text)